fix: Authorities object-fit, reduce space between title and texton Tree
